test: generalize the d.ts export-coverage guard to core + server (#388)
Extends the drift guard to every published entry point with a hand-maintained
.d.ts overlay (@webjsdev/core, @webjsdev/server, @webjsdev/server/testing),
each reading its runtime exports dynamically and tsc-checking that the overlay
declares them all. Server was already complete (the 12 audited helpers are
correctly scoped to /testing with types); this locks all three against future
drift.
